Moved EnergyMode into types.py, added new imports
