+++ src/dodal/devices/electron_analyser/__init__.py
@@ -1,7 +1,8 @@
-from .specs import SpecsDetector
+from .types import EnergyMode
 from .util import to_binding_energy, to_kinetic_energy
-from .vgscienta import VGScientaDetector
 
-TElectronAnalyserDetectorImpl = VGScientaDetector | SpecsDetector
-
-__all__ = ["to_binding_energy", "to_kinetic_energy", "TElectronAnalyserDetectorImpl"]
+__all__ = [
+    "to_binding_energy",
+    "to_kinetic_energy",
+    "EnergyMode",
+]
