+++ src/dodal/devices/electron_analyser/abstract/base_driver_io.py
@@ -13,10 +13,8 @@ from ophyd_async.core import (
 from ophyd_async.epics.adcore import ADBaseIO
 from ophyd_async.epics.core import epics_signal_r, epics_signal_rw
 
-from dodal.devices.electron_analyser.abstract.base_region import (
-    EnergyMode,
-    TAbstractBaseSequence,
-)
+from dodal.devices.electron_analyser.abstract.base_region import TAbstractBaseSequence
+from dodal.devices.electron_analyser.types import EnergyMode
 from dodal.devices.electron_analyser.util import to_binding_energy
