+++ src/dodal/devices/electron_analyser/util.py
@@ -1,4 +1,4 @@
-from dodal.devices.electron_analyser.abstract.base_region import EnergyMode
+from dodal.devices.electron_analyser.types import EnergyMode
 
 
 def to_kinetic_energy(
