+++ tests/devices/unit_tests/electron_analyser/vgscienta/test_region.py
@@ -3,7 +3,8 @@ from typing import Any
 import pytest
 
 from dodal.common.data_util import load_json_file_to_class
-from dodal.devices.electron_analyser.abstract import EnergyMode, TAbstractBaseRegion
+from dodal.devices.electron_analyser.abstract import TAbstractBaseRegion
+from dodal.devices.electron_analyser.types import EnergyMode
 from dodal.devices.electron_analyser.vgscienta import VGScientaRegion, VGScientaSequence
 from dodal.devices.electron_analyser.vgscienta.region import (
     AcquisitionMode,
